Make less resumey and add TARR
Really, a lot of stuff on there was probably a bit too personal (and it
is now forever presered in this git repo and probably on the internet
archive, but, anyway), and really was kind of resume level stuff, rather
than just public internet me level stuff.

So, I've basically cut it down to just projects, and in doing so also
added my must recent project, The Anthropocene Reviewed Reviewed.